requirements.txt
segmentation_models_pytorch
numpy 
scikit-learn 
matplotlib 
pandas
torchmetrics
wandb
torchinfo
albumentations